examples/hyperdatasets/requirements.txt
allegroai~=3.5.7
clearml
numpy
Pillow
torch
